Large CodeQL remediation batch
- apply broad automated remediation passes across portal, WebDAV, and shared code
- keep semantic fixes that compiled cleanly after regression repair
- repair automation regressions in portal adapters, helpers, Exchange screens, and shared utilities
- validate with successful FuseCP.Portal.Modules build

+++ FuseCP/Sources/FuseCP.WebPortal/DesktopModules/FuseCP/UserControls/ScheduleInterval.ascx.cs
@@ -55,19 +55,19 @@ namespace FuseCP.Portal
                 {
                     // days
                     ddlUnits.SelectedIndex = 0;
-                    txtInterval.Text = ((int)(s / 86400)).ToString();
+                    txtInterval.Text = ((s / 86400)).ToString();
                 }
                 else if (s % 3600 == 0)
                 {
                     // hours
                     ddlUnits.SelectedIndex = 1;
-                    txtInterval.Text = ((int)(s / 3600)).ToString();
+                    txtInterval.Text = ((s / 3600)).ToString();
                 }
                 else if (s % 60 == 0)
                 {
                     // minutes
                     ddlUnits.SelectedIndex = 2;
-                    txtInterval.Text = ((int)(s / 60)).ToString();
+                    txtInterval.Text = ((s / 60)).ToString();
                 }
                 else
                 {
